Created an array of objects to be destructured into reusable card component for new additions to the project list.
Created an array of objects to be destructured into reusable card component for new additions to the project list. I only need to add a new object to projectList[_] to add a new project. The card is already created. Next I need to align the cards and resize them. Also thinking of adding quick scrolling to the next set of project cards.